+++ source/extensions/omni.flux.stage_manager.plugin.interaction.usd/omni/flux/stage_manager/plugin/interaction/usd/all_lights.py
@@ -42,7 +42,7 @@ class AllLightsInteractionPlugin(_StageManagerUSDInteractionPlugin):
         "CustomTagsWidgetPlugin",
         "FocusInViewportActionWidgetPlugin",
         "IsCaptureStateWidgetPlugin",
+        "IsCategoryHiddenStateWidgetPlugin",
         "IsVisibleActionWidgetPlugin",
         "PrimTreeWidgetPlugin",
-        "IsCategoryHiddenStateWidgetPlugin",
     ]