+++ source/extensions/omni.flux.validator.plugin.context.usd_stage/omni/flux/validator/plugin/context/usd_stage/asset_importer.py
@@ -34,6 +34,7 @@ from omni.flux.asset_importer.core.data_models import UsdExtensions as _UsdExten
 from omni.flux.asset_importer.widget.file_import_list import FileImportListModel as _FileImportListModel
 from omni.flux.asset_importer.widget.file_import_list import FileImportListWidget as _FileImportListWidget
 from omni.flux.info_icon.widget import InfoIconWidget as _InfoIconWidget
+from omni.flux.utils.common.api import send_request as _send_request
 from omni.flux.utils.common.decorators import ignore_function_decorator as _ignore_function_decorator
 from omni.flux.utils.common.omni_url import OmniUrl as _OmniUrl
 from omni.flux.utils.widget.file_pickers import open_file_picker as _open_file_picker
@@ -64,6 +65,7 @@ class AssetImporter(_ContextBaseUSD):
         full_path_keep: bool = False
         full_path_root: Optional[_OmniUrl] = None
         close_stage_on_exit: bool = False
+        default_output_endpoint: Optional[str] = None  # An API endpoint to hit up to get the default output directory
 
         _compatible_data_flow_names = ["InOutData"]
         data_flows: Optional[List[_InOutDataFlow]] = None  # override base argument with the good typing
@@ -481,6 +483,15 @@ class AssetImporter(_ContextBaseUSD):
                     mouse_pressed_fn=partial(self.__open_dialog, schema_data, self._output_field.model),
                 )
                 _InfoIconWidget("The directory to import the converted input files to.")
+
+            if schema_data.default_output_endpoint:
+                try:
+                    response = await _send_request("GET", schema_data.default_output_endpoint)
+                    output_directory = response.get("asset_path")
+
+                    schema_data.output_directory = _OmniUrl(output_directory)
+                except RuntimeError:
+                    pass
 
             self._output_field.model.set_value(
                 carb.tokens.get_tokens_interface().resolve(str(schema_data.output_directory))