+++ source/extensions/lightspeed.trex.asset_replacements.core.shared/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [2.5.0]
+## Changed
+- Changed `get_default_output_directory_with_data_model` to add the ability to specify which default path to use
+
 ## [2.4.0]
 ## Changed
 - Separated skeleton logic into its own module