+++ source/extensions/lightspeed.trex.stage_manager.plugin.interaction.usd/docs/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.1.1]
+### Fixed
+- Added missing compatible widgets
+- Fixed tooltip and compatible trees
+
 ## [1.1.0]
 ### Added
 - Added stage manager mesh tab