+++ source/extensions/omni.flux.tests.settings/docs/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.2.0]
+### Added
+- Added Default address and port for `omni.services.transport.server.http`
+
 ## [1.1.0]
 ### Added
 - Added Read-Only credentials for the Nucleus Registry